Organize default story plugins better.
