+++ packages/apps/plugins/plugin-dnd/src/stories/DndPluginDefaultStoryPlugin.tsx
@@ -1,0 +1,80 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import { useSortable } from '@dnd-kit/sortable';
+import { CSS } from '@dnd-kit/utilities';
+import { faker } from '@faker-js/faker';
+import { DotsSixVertical } from '@phosphor-icons/react';
+import React, { createContext, PropsWithChildren, useContext } from 'react';
+
+import { randomString } from '@dxos/aurora';
+import { mx } from '@dxos/aurora-theme';
+import { createStore } from '@dxos/observable-object';
+import { Surface } from '@dxos/react-surface';
+
+type StoryItemProps = { id: string; title: string };
+
+faker.seed(1111);
+
+const defaultItems = {
+  items: [
+    { id: `storyItem:${randomString()}`, title: faker.commerce.product() },
+    { id: `storyItem:${randomString()}`, title: faker.commerce.product() },
+    { id: `storyItem:${randomString()}`, title: faker.commerce.product() },
+    { id: `storyItem:${randomString()}`, title: faker.commerce.product() },
+  ],
+};
+
+export const StoryItemDragOverlay = ({ data }: { data: string }) => {
+  // todo(thure): the `store` here is not the one set by the provider, why is that?
+  const store = useContext(DndPluginStoryPluginContext);
+  const item = store.items.find(({ id }) => id === data);
+  return item ? <StoryItem {...item} dragOverlay /> : null;
+};
+
+export const StoryItem = ({ id, title, dragging }: StoryItemProps & { dragOverlay?: boolean; dragging?: boolean }) => {
+  const { attributes, listeners, setNodeRef, transform, transition } = useSortable({
+    id,
+  });
+  return (
+    <div
+      className={mx('p-2 flex gap-2 items-center', dragging && 'invisible')}
+      style={{ transform: CSS.Translate.toString(transform), transition }}
+      {...attributes}
+      {...listeners}
+      ref={setNodeRef}
+    >
+      <DotsSixVertical />
+      {title}
+    </div>
+  );
+};
+
+export const DndPluginStoryPluginContext = createContext<{ items: StoryItemProps[] }>(defaultItems);
+
+const DndPluginDefaultStoryPluginDefault = () => {
+  return (
+    <div role='none' className='flex p-4 gap-4'>
+      <Surface role='dndpluginstory' />
+    </div>
+  );
+};
+
+export const DndPluginDefaultStoryPlugin = () => {
+  const store = createStore<{ items: StoryItemProps[] }>(defaultItems);
+  return {
+    meta: {
+      id: 'dxos:dndStoryPluginA',
+    },
+    provides: {
+      context: ({ children }: PropsWithChildren) => (
+        <DndPluginStoryPluginContext.Provider value={store}>{children}</DndPluginStoryPluginContext.Provider>
+      ),
+      components: {
+        default: DndPluginDefaultStoryPluginDefault,
+      },
+      dndStory: store,
+    },
+  };
+};
